+++ src/zns.ts
@@ -119,7 +119,7 @@ export default class Zns extends NamingService {
    */
   async address(domain: string, currencyTicker: string): Promise<string> {
     const data = await this.resolve(domain);
-    if ((data && !data.meta.owner )|| isNullAddress(data!.meta.owner))
+    if ((data && !data.meta.owner) || isNullAddress(data!.meta.owner))
       throw new ResolutionError(ResolutionErrorCode.UnregisteredDomain, {
         domain,
       });
@@ -187,7 +187,7 @@ export default class Zns extends NamingService {
    */
   isSupportedDomain(domain: string): boolean {
     const tokens = domain.split('.');
-    return (!!tokens.length && tokens[tokens.length - 1] === 'zil' );
+    return !!tokens.length && tokens[tokens.length - 1] === 'zil';
   }
 
   /**