+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Using flexible dependacies instead of locked versions
 * Moved sizecheck to a separate dev dependacy
 * Added web3Provider Support [#57]
+* Added factories Resolution.infura, Resolution.provider, Resolution.jsonRPCprovider 
 
 ## 1.0.24
 * Bug fix, namehash the domain before asking for a resolver on cns.